calculateBatchThreads: API changes, bugfix
There was an issue with the HGW calculation if hackDifficulty was below
~8.57. This is fixed now, and all the calculations should work for for
all valid ranges. However, they are still lacking some input/output
validation, so it is still possible to get negative or excessive numbers
of threads if you give outlandish enough options. (For instance, if you
specify a fixed number of grow threads that is greater than what is
needed to grow to full, currently you will get hack threads greater than
what is needed to drain 100%.)

In preparation for input/output validation and clamping there is an API
breaking change: the return value is now {hack, grow} instead of a
single number.

Also, the "fortify" parameter has changed name to "weaken", because that
is what it does. ("Fortify" goes in the wrong direction.)